requirements.txt
amqp==2.2.2
anyjson==0.3.3
arrow==0.12.1
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
Django==2.0.2
django-celery==3.2.2
django-cors-headers==2.2.0
djangorestframework==3.7.7
flake8==3.5.0
flake8-polyfill==1.0.2
idna==2.5
mccabe==0.6.1
pep8-naming==0.5.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.10.0
urllib3==1.21.1
vine==1.1.4
virtualenv==15.1.0
yamllint==1.7.0
